# api/docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_logic3579_optimus_internal_infra_response.Envelope:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      message_key:
        type: string
    type: object
...
  internal_modules_menu.CreateRequest:
    properties:
      code:
        maxLength: 64
        minLength: 2
        type: string
      component:
        maxLength: 255
        type: string
      hidden:
        type: boolean
      icon:
        maxLength: 64
        type: string
      name:
        maxLength: 128
        type: string
      parent_id:
        type: integer
      path:
        maxLength: 255
        type: string
      permission_code:
        maxLength: 128
        type: string
      sort_order:
        type: integer
    required:
    - code
    - name
    type: object
  internal_modules_menu.UpdateRequest:
    properties:
      component:
        maxLength: 255
        type: string
      hidden:
        type: boolean
      icon:
        maxLength: 64
        type: string
      name:
        maxLength: 128
        type: string
      parent_id:
        type: integer
      path:
        maxLength: 255
        type: string
      permission_code:
        maxLength: 128
        type: string
      sort_order:
        type: integer
    type: object
...
  internal_modules_observability_dashboard.Detail:
    properties:
      created_at:
        type: string
      created_by_user_id:
        type: integer
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      panels:
        items:
          $ref: '#/definitions/internal_modules_observability_dashboard.Panel'
        type: array
      refresh_interval_s:
        type: integer
      time_range:
        type: string
      updated_at:
        type: string
    type: object
  internal_modules_observability_dashboard.ListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/internal_modules_observability_dashboard.Detail'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  internal_modules_observability_dashboard.Panel:
    properties:
      created_at:
        type: string
      dashboard_id:
        type: integer
      datasource_id:
        type: integer
      id:
        type: integer
      legend:
        type: string
      panel_type:
        type: string
      promql:
        type: string
      sort_order:
        type: integer
      title:
        type: string
      unit:
        type: string
      updated_at:
        type: string
      width:
        type: integer
    type: object
  internal_modules_observability_dashboard.PanelInput:
    properties:
      datasource_id:
        type: integer
      legend:
        maxLength: 128
        type: string
      panel_type:
        type: string
      promql:
        maxLength: 8192
        type: string
      sort_order:
        type: integer
      title:
        maxLength: 128
        type: string
      unit:
        type: string
      width:
        type: integer
    required:
    - datasource_id
    - panel_type
    - promql
    - title
    - unit
    - width
    type: object
  internal_modules_observability_dashboard.SaveRequest:
    properties:
      description:
        maxLength: 4096
        type: string
      name:
        maxLength: 128
        type: string
      panels:
        items:
          $ref: '#/definitions/internal_modules_observability_dashboard.PanelInput'
        type: array
      refresh_interval_s:
        type: integer
      time_range:
        type: string
    required:
    - name
    - panels
    - refresh_interval_s
    - time_range
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: |-
    P0 admin backend for Optimus — auth, RBAC, users, roles, permissions, menus, audit.
    All authenticated endpoints expect `Authorization: Bearer <access_token>`.
  title: Optimus Admin API
  version: "1.0"
paths:
...
  /me/menus:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_logic3579_optimus_internal_infra_response.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_logic3579_optimus_internal_infra_response.Envelope'
      security:
      - BearerAuth: []
      summary: Current user menu tree
      tags:
      - me
...
  /menus:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_logic3579_optimus_internal_infra_response.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_logic3579_optimus_internal_infra_response.Envelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_logic3579_optimus_internal_infra_response.Envelope'
      security:
      - BearerAuth: []
      summary: Menu tree
      tags:
      - menus
    post:
      consumes:
      - application/json
      parameters:
      - description: menu payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_menu.CreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_logic3579_optimus_internal_infra_response.Envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_logic3579_optimus_internal_infra_response.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_logic3579_optimus_internal_infra_response.Envelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_logic3579_optimus_internal_infra_response.Envelope'
      security:
      - BearerAuth: []
      summary: Create menu
      tags:
      - menus
  /menus/{id}:
    delete:
      parameters:
      - description: menu ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_logic3579_optimus_internal_infra_response.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_logic3579_optimus_internal_infra_response.Envelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_logic3579_optimus_internal_infra_response.Envelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_logic3579_optimus_internal_infra_response.Envelope'
      security:
      - BearerAuth: []
      summary: Delete menu
      tags:
      - menus
    put:
      consumes:
      - application/json
      parameters:
      - description: menu ID
        in: path
        name: id
        required: true
        type: integer
      - description: menu payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_menu.UpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_logic3579_optimus_internal_infra_response.Envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_logic3579_optimus_internal_infra_response.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_logic3579_optimus_internal_infra_response.Envelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_logic3579_optimus_internal_infra_response.Envelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_logic3579_optimus_internal_infra_response.Envelope'
      security:
      - BearerAuth: []
      summary: Update menu
      tags:
      - menus
...
  /observability/dashboards:
    get:
      parameters:
      - description: name search
        in: query
        name: q
        type: string
      - default: 1
        description: page
        in: query
        name: page
        type: integer
      - default: 20
        description: page size
        in: query
        name: page_size
        type: integer
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_logic3579_optimus_internal_infra_response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/internal_modules_observability_dashboard.ListResponse'
              type: object
      security:
      - BearerAuth: []
      summary: List custom metric dashboards
      tags:
      - observability
    post:
      consumes:
      - application/json
      parameters:
      - description: dashboard aggregate
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_observability_dashboard.SaveRequest'
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_logic3579_optimus_internal_infra_response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/internal_modules_observability_dashboard.Detail'
              type: object
      security:
      - BearerAuth: []
      summary: Create a custom metric dashboard
      tags:
      - observability
  /observability/dashboards/{id}:
    delete:
      parameters:
      - description: dashboard ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_logic3579_optimus_internal_infra_response.Envelope'
      security:
      - BearerAuth: []
      summary: Delete a custom metric dashboard
      tags:
      - observability
    get:
      parameters:
      - description: dashboard ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_logic3579_optimus_internal_infra_response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/internal_modules_observability_dashboard.Detail'
              type: object
      security:
      - BearerAuth: []
      summary: Get a custom metric dashboard
      tags:
      - observability
    put:
      consumes:
      - application/json
      parameters:
      - description: dashboard ID
        in: path
        name: id
        required: true
        type: integer
      - description: dashboard aggregate
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_observability_dashboard.SaveRequest'
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_logic3579_optimus_internal_infra_response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/internal_modules_observability_dashboard.Detail'
              type: object
      security:
      - BearerAuth: []
      summary: Replace a custom metric dashboard
      tags:
      - observability
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the JWT access token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"