# internal/modules/menu/handler.go
package menu

import (
	"strconv"

	"github.com/gin-gonic/gin"

	apperr "github.com/logic3579/optimus/internal/infra/errors"
	"github.com/logic3579/optimus/internal/infra/middleware"
	"github.com/logic3579/optimus/internal/infra/response"
)

type Handler struct {
	svc *Service
}

func NewHandler(svc *Service) *Handler { return &Handler{svc: svc} }

func (h *Handler) Register(g *gin.RouterGroup) {
	g.GET("", h.tree)
	g.POST("", h.create)
	g.PUT("/:id", h.update)
	g.DELETE("/:id", h.delete)
}

// Public wrappers used by main.go to register handlers under nested groups
// gated by middleware.RequirePermission.
func (h *Handler) HandleTree() gin.HandlerFunc   { return h.tree }
func (h *Handler) HandleCreate() gin.HandlerFunc { return h.create }
func (h *Handler) HandleUpdate() gin.HandlerFunc { return h.update }
func (h *Handler) HandleDelete() gin.HandlerFunc { return h.delete }

func (h *Handler) actor(c *gin.Context) uint64 { return c.GetUint64(middleware.CtxKeyUserID) }

func (h *Handler) parseID(c *gin.Context) (uint64, bool) {
	id, err := strconv.ParseUint(c.Param("id"), 10, 64)
	if err != nil || id == 0 {
		response.Error(c, apperr.New(apperr.CodeBadRequest, "common.bad_request", "invalid id"))
		return 0, false
	}
	return id, true
}

// tree returns the full menu tree.
// @Summary  Menu tree
// @Tags     menus
// @Security BearerAuth
// @Produce  json
// @Success  200 {object} response.Envelope
// @Failure  401 {object} response.Envelope
// @Failure  403 {object} response.Envelope
// @Router   /menus [get]
func (h *Handler) tree(c *gin.Context) {
	out, err := h.svc.Tree(c.Request.Context())
	if err != nil {
		response.Error(c, err)
		return
	}
	response.Success(c, out)
}

// create creates a menu node.
// @Summary  Create menu
// @Tags     menus
// @Security BearerAuth
// @Accept   json
// @Produce  json
// @Param    body body     CreateRequest true "menu payload"
// @Success  200  {object} response.Envelope
// @Failure  400  {object} response.Envelope
// @Failure  401  {object} response.Envelope
// @Failure  403  {object} response.Envelope
// @Router   /menus [post]
func (h *Handler) create(c *gin.Context) {
	var req CreateRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		response.Error(c, apperr.New(apperr.CodeValidation, "common.validation", err.Error()))
		return
	}
	out, err := h.svc.Create(c.Request.Context(), h.actor(c), c.ClientIP(), c.Request.UserAgent(), req)
	if err != nil {
		response.Error(c, err)
		return
	}
	response.Success(c, out)
}

// update mutates a menu node.
// @Summary  Update menu
// @Tags     menus
// @Security BearerAuth
// @Accept   json
// @Produce  json
// @Param    id   path     int           true "menu ID"
// @Param    body body     UpdateRequest true "menu payload"
// @Success  200  {object} response.Envelope
// @Failure  400  {object} response.Envelope
// @Failure  401  {object} response.Envelope
// @Failure  403  {object} response.Envelope
// @Failure  404  {object} response.Envelope
// @Router   /menus/{id} [put]
func (h *Handler) update(c *gin.Context) {
	id, ok := h.parseID(c)
	if !ok {
		return
	}
	var req UpdateRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		response.Error(c, apperr.New(apperr.CodeValidation, "common.validation", err.Error()))
		return
	}
	out, err := h.svc.Update(c.Request.Context(), h.actor(c), c.ClientIP(), c.Request.UserAgent(), id, req)
	if err != nil {
		response.Error(c, err)
		return
	}
	response.Success(c, out)
}

// delete removes a menu node.
// @Summary  Delete menu
// @Tags     menus
// @Security BearerAuth
// @Produce  json
// @Param    id  path     int true "menu ID"
// @Success  200 {object} response.Envelope
// @Failure  401 {object} response.Envelope
// @Failure  403 {object} response.Envelope
// @Failure  404 {object} response.Envelope
// @Router   /menus/{id} [delete]
func (h *Handler) delete(c *gin.Context) {
	id, ok := h.parseID(c)
	if !ok {
		return
	}
	if err := h.svc.Delete(c.Request.Context(), h.actor(c), c.ClientIP(), c.Request.UserAgent(), id); err != nil {
		response.Error(c, err)
		return
	}
	response.Success(c, gin.H{"ok": true})
}